# v1.1/backend/docs/swagger.yaml
definitions:
...
  common.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
...
  iTask.TaskRequest:
    properties:
      direction:
        description: 扩展方向（可选，如：left/right/top/bottom）
        type: string
      expandRatio:
        description: 扩展比例（可选）
        type: number
      imageUrl:
        description: 图片URL（必需，需验证URL格式）
        type: string
      prompt:
        description: 用户提示词（必需）
        type: string
    required:
    - imageUrl
    - prompt
    type: object
  image_expand_task.ITaskVO:
    properties:
      aiRecap:
        description: AI说明（可能含敏感信息）
        type: string
      createTime:
        description: 【重要】VO建议增加时间字段
        type: string
      execMessage:
        description: 执行错误时才返回
        type: string
      expandParams:
        description: 参数配置（内部用可脱敏）
        type: string
      expandedPicUrl:
        description: 结果图URL（未生成时隐藏）
        type: string
      id:
        description: ID保持原样
        example: ""
        type: string
      name:
        description: 任务名称
        type: string
      originalPicUrl:
        description: 原图URL（可考虑返回脱敏后的CDN地址）
        type: string
      pictureId:
        description: 关联图片ID（非必返）
        example: ""
        type: string
      prompt:
        description: 【脱敏】用户提示词（敏感时可omitempty）
        type: string
      status:
        description: 任务状态（必须返回）
        type: string
    type: object
...
host: localhost:8001
info:
  contact:
    email: lwh24621@qq.com
    name: lwhhhh
    url: 无
  description: CanvasCloud
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: 无服务条款
  title: CanvasCloud
  version: "1.0"
paths:
  /v1/file/test/download:
    get:
      parameters:
      - description: 文件存储在 COS 的 KEY
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 返回文件流
          schema:
            type: file
        "400":
          description: 下载失败，详情见响应中的 code
          schema:
            $ref: '#/definitions/common.Response'
      summary: 测试文件下载接口「管理员」
      tags:
      - file
  /v1/file/test/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 响应文件存储在COS的KEY
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      summary: 测试文件上传接口「管理员」
      tags:
      - file
...
  /v1/picture/out_painting/delete_task/:id:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功，返回图片信息视图
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 删除任务视图「需要登录校验」
      tags:
      - picture
  /v1/picture/out_painting/list:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回图片信息视图
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/image_expand_task.ITaskVO'
                  type: array
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取ai请求任务视图「需要登录校验」
      tags:
      - picture
  /v1/picture/out_painting/procreate_task:
    post:
      consumes:
      - application/json
      parameters:
      - description: 任务
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/iTask.TaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功，返回任务信息视图
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: 更新失败，详情见响应中的code
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 上传ai扩图图片任务PRO「需要登录校验」
      tags:
      - picture
...
securityDefinitions:
  BearerAuth:
    description: 'JWT 认证格式: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"